Renamed headers to include, changed every size_t to unsigned long to stop needing to include stddef.h

+++ include/Comment.h
@@ -5,18 +5,16 @@
 #ifndef REDDIT_COMMENT_H
 #define REDDIT_COMMENT_H
 
-#include <defs.h>
-
 typedef struct Comment {
     const char* title;
     const char* author;
     const char* body;
-    size_t score;
+    unsigned long score;
     const char* id;
     const char* thumbnail;
     const char* url;
     struct Comment** children;
-    size_t no_children;
+    unsigned long no_children;
 } Comment_t;
 
 typedef void (*comment_cb)(Comment_t*, void*, int);
