+++ include/List.h
@@ -5,8 +5,6 @@
 #ifndef LIBREDDIT_LIST_H
 #define LIBREDDIT_LIST_H
 
-#include <defs.h>
-
 #ifdef __cplusplus
 extern "C"{
 #endif
@@ -15,14 +13,14 @@ typedef void(* freer_cb)(void*);
 
 typedef struct List{
     void** data;
-    size_t length;
+    unsigned long length;
     freer_cb freer;
 #ifdef PRIVATE
-    size_t elem_size;
+    unsigned long elem_size;
 #endif
 } List_t;
 
-List_t* list_new(size_t elem_size, freer_cb freer);
+List_t* list_new(unsigned long elem_size, freer_cb freer);
 
 void list_append(List_t *lst, void* data);
 
@@ -30,7 +28,7 @@ void list_free(List_t* lst);
 
 #define LIST(type, freer) list_new(sizeof(type), freer)
 
-#define LIST_FREE(lst, type) for(size_t i=0 ; i<lst->length ; i++){ lst->freer((type *)lst->data[i]); }
+#define LIST_FREE(lst, type) for(unsigned long i=0 ; i<lst->length ; i++){ lst->freer((type *)lst->data[i]); }
 
 #ifdef __cplusplus
 }
